pyproject.toml
[project]
name = "egm-state-detector"
version = "0.1.0"
description = "EGM Stream State Detector with ROI/ImageHash"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
    "Pillow>=9.0",
    "imagehash>=4.3",
    "PyYAML>=6.0",
    "pydantic>=2.0",
]
requires-python = ">=3.9"

[project.scripts]
egm-detector = "egm_state_detector.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
